requirements.txt
imageio==2.9.0
jax==0.2.8
jaxlib==0.1.59
jupyter==1.0.0
matplotlib==3.3.3
numpy==1.18
pandas==1.1.4
python-igraph==0.8.3
tqdm==4.46.0
